requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
pandas==2.2.1
scikit-learn==1.4.1.post1
requests==2.31.0
mysql-connector-python